refactor(core): own provider identity tokens
