+++ docs/topology-status.md
@@ -25,13 +25,13 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-semantic | — | 6 | 6 | 0 | 0 |
 | archive-session | — | 6 | 6 | 0 | 0 |
 | archive-viewport | — | 5 | 5 | 0 | 0 |
-| core-primitive | — | 9 | 9 | 0 | 0 |
+| core-primitive | — | 10 | 10 | 0 | 0 |
 
 ### Summary
 
 - **Stable** (no move scoped): 582
 - **Kernel** (polylogue/ root): 10
-- **Primitives** (lib-root + storage-root): 14
+- **Primitives** (lib-root + storage-root): 13
 - **TBD** (cell needs explicit assignment): 0
 - **Total declared**: 695
 - **Realized polylogue/**/*.py**: 695 files declared
