+++ polylogue/schemas/runtime_registry.py
@@ -12,8 +12,8 @@ from pathlib import Path
 
 from polylogue.archive.raw_payload.decode import JSONRecord
 from polylogue.core.json import json_document
-from polylogue.lib.provider_identity import canonical_schema_provider as _canonical_schema_provider
-from polylogue.lib.provider_identity import normalize_provider_token
+from polylogue.core.provider_identity import canonical_schema_provider as _canonical_schema_provider
+from polylogue.core.provider_identity import normalize_provider_token
 from polylogue.paths import data_home
 from polylogue.schemas.observation import (
     derive_bundle_scope,
