+++ polylogue/schemas/sampling_db.py
@@ -11,7 +11,7 @@ from typing import Literal, TypeAlias, overload
 from polylogue.archive.raw_payload import extract_record_samples_from_raw_content
 from polylogue.archive.raw_payload.decode import RawPayloadEnvelope
 from polylogue.core.json import JSONDocument
-from polylogue.lib.provider_identity import (
+from polylogue.core.provider_identity import (
     CORE_RUNTIME_PROVIDERS,
     canonical_runtime_provider,
     canonical_schema_provider,
