+++ tests/unit/core/test_models.py
@@ -17,11 +17,11 @@ from polylogue.archive.message.roles import Role
 from polylogue.archive.raw_payload import build_raw_payload_envelope
 from polylogue.archive.viewport.viewports import ToolCall, classify_tool
 from polylogue.core.json import JSONDocument, JSONValue, json_document
-from polylogue.lib.models import Message
-from polylogue.lib.provider_identity import (
+from polylogue.core.provider_identity import (
     canonical_runtime_provider,
     canonical_schema_provider,
 )
+from polylogue.lib.models import Message
 from polylogue.schemas.code_detection.detection import LANGUAGE_PATTERNS, detect_language, extract_code_block
 from polylogue.storage.hydrators import (
     attachment_from_record,
